==> 15-following-network.Rmd <==


rm(list = ls())

library(igraph)
library(Matrix)
library(irlba)
library(ggplot2)
library(data.table)
source("function.R")



######
```{r}
load("../data/follower_Ne")


load("../data/followers_Network/data.RData")
library(igraph)
library(Matrix)
library(irlba)
library(ggplot2)
library(data.table)
source("function.R")

name1 <- rownames(A)
name2 <- colnames(A)
deg_row = rowSums(A); deg_col = colSums(A)

summary(count2)
summary(deg_col)

#
hist(log(1+deg_row), breaks = 100)
hist(log(1+deg_col), breaks = 100)
#hist(log(1+count1), breaks = 100)
hist(log(1+count2), breaks = 100)

#id.trump = which.max(deg_col)
#A <- A[ ,-id.trump]

n = dim(A)[1]
m = dim(A)[2]


## using population col deg + inner product
A1 <- A %*% Diagonal(m, (deg_col/ count2) * log( (n+1)/(deg_col+1) )  ) # min(count2)>=1000, can be ommitted
norm1 <- rowSums(A1 *A1);     norm1 <- sqrt(norm1 + 1e-4)                          #apply cannot work, saying too large

L =  Diagonal(n, norm1^(-0.5)) %*% A1
svd_L <- irlba(L, nv = 50)





plot(svd_L$d)


#normalization of rows
k = 50 # or 7
U <- svd_L$u[,1:k] %*%Diagonal(k, svd_L$d[1:k]^(1/2))      # first eigenvectors are more important??

#km_row1 = kmeans(U, k, nstart = 500, iter.max =50)

rowN <- rowSums(U*U); rowN <- sqrt(rowN +1e-6)
U1 <- Diagonal(length(rowN), rowN^(-1))%*%U
set.seed(123)
km_row2 = kmeans(U1, k, nstart = 100, iter.max =50)



km_row = km_row2
# representative words
top = 40
keyfriends <- matrix("", top, k)
scores <- matrix(0, top, k)
for(i in 1:k){
  c_i1 <- colMeans( L[which(km_row$cluster==i), ] )
  c_i2 <- colMeans( L[which(km_row$cluster!=i), ] )
  #variance stablization transformation
  c_i <- sqrt(c_i1) - sqrt(c_i2)   
  names_tmp <- colnames(A)[order(-c_i)]
  c_i <- c_i[order(-c_i)]
  idx <- which(!is.na(names_tmp))[1:top]  
  
  keyfriends[,i]<- names_tmp[idx] 
  scores[,i] <- round(c_i[idx],3)
  print(c_i['realDonaldTrump'])
}



library(smappR)
keyfriends_info <- getUsersBatch( 
  screen_names = as.vector(keyfriends),                                
  include_entities = T, 
  oauth_folder = "./credentials/credential_mixed/")


clustering <- cbind(as.vector(keyfriends),rep(1:k, each = top),
                    rep(km_row$size, each = top),  as.vector(scores))
clustering <- data.frame(clustering, stringsAsFactors = F)
names(clustering) <- c("screen_names","clusters","Sizes","scores")
keyfriends_info <- data.table(keyfriends_info)
setkey(keyfriends_info, screen_name)  #order by screen_name, fast match
result <- cbind(clustering,keyfriends_info[clustering$screen_name])
result$description <- gsub("[\t\n\r]", " ", result$description)
result1 <- result[,c(1,8,13,14,2,4,3,5,7,9,10,11,12,15,16:18)]
write.csv(result, file ="./1209/following/k50/distinctive_friends.csv", row.names = F)


sn_cluster <- data.frame(cbind( screenNames = rownames(A), cluster = km_row$cluster))
ids <- ID_SNs$id_str[match(sn_cluster$screenNames,ID_SNs$screen_name)]
sn_cluster <- cbind(id_str = ids, sn_cluster)
write.csv(sn_cluster, file = "./1209/following/k50/id_sn_cluster.csv", row.names = F)

#  2-dimension visualization:
## bloomplot 
Z <- matrix(0, n, k )
for ( i in 1:k){
  Z[which(km_row$cluster == i), i ] <- 1
}
NZ <- Z %*% Diagonal(k, colSums(Z)^(-1))

B1 <- t(NZ) %*% A1 %*%t(A1) %*% NZ; image(B1)
B2 <- t(NZ) %*% L %*%t(L) %*% NZ; image(B2)


pdf("./1209/following/k50/blockB.pdf", height = 7, width = 8)
ggplot(data = melt(as.matrix(sqrt(B1))), aes(x=Var1, y =Var2, fill = value))+
  geom_tile() + labs(title= expression(sqrt(B1)))+xlab("row")+ylab("col") 
ggplot(data = melt(as.matrix(sqrt(B2))), aes(x=Var1, y =Var2, fill = value))+
  geom_tile() + labs(title= expression(sqrt(B2)))+xlab("row")+ylab("col") 
dev.off()
colSums(Z);colSums(Y)


confMatrix <- t(Z1) %*% Y
pdf(file = "./1209/following/k50/blockB.pdf", onefile = T, width = 8, height = 7)
library(ggplot2)
confMatrix1 <- Diagonal(dim(Z1)[2], rowSums(confMatrix)^(-1))%*% confMatrix;
ggplot(melt(as.matrix(confMatrix1)), aes(x=Var1, y= Var2, fill=value)) + 
  geom_tile()+ labs(title = "confMat1") + xlab("row") + ylab("col")
confMatrix2 <- confMatrix %*% Diagonal(dim(Y)[2], colSums(confMatrix)^(-1));
ggplot(melt(as.matrix(confMatrix2)),aes(x=Var1, y=Var2, fill=value)) + 
  geom_tile()+ labs(title = "confMat2") + xlab("row") + ylab("col")
confMatrix3 <- Diagonal(dim(Z1)[2], rowSums(confMatrix)^(-1))%*% confMatrix %*% Diagonal(dim(Y)[2], colSums(confMatrix)^(-1))
ggplot(melt(as.matrix(confMatrix3)),aes(x=Var1, y=Var2, fill=value)) + 
  geom_tile()+ labs(title = "confMat3") + xlab("row") + ylab("col") #+scale_fill_gradient(low="green", high="red")
dev.off()
NMI(confMatrix)








set.seed(12)
samp1 <- sample(1:nrow(U), 2000)
pdf("../data/followers_Network/LL3_k7_n.pdf")
plot(U1[samp1,2],U1[samp1,3], pch =km_row$cluster[samp1],  col = as.factor(km_row$cluster[samp1]))
legend("topright", legend = paste0(1:k,"-",km_row$size), pch=1:k, col =as.factor(1:k) )
dev.off()




weight = (deg_col/ count2)
A1 <- A %*% Diagonal(m, weight)
l1 <- rowSums(A1*A1); l1 <- sqrt(l1+1e-6); A2 <- Diagonal(n, l1^(-1))
nnzeros <- colSums(A>0)
A3 <- A2 * A1 * log( (n+1)/(nnzeros+1) )   # min(count2)>=1000, can be ommitted


L4 <- A %*% Diagonal(m, sqrt(deg_col/ count2) * log( (n+1)/(deg_col+1) )  ) # min(count2)>=1000, can be ommitted
norm1 <- rowSums(L4 *L4)                               #apply cannot work, saying too large
norm1 <- norm1 + 0.1*mean(norm1) 
L4 =  Diagonal(n, norm1^(-0.5)) %*% L4
svd_L4 <- irlba(L4, nv = 50)
