==> 16-retweet-network-allTweets-combined.Rmd <==
---
title: "Analysis on Followers' Interaction with Trump on Twitter"
author: "Song Wang"
date: "`r Sys.Date()`"
output: html_document
---



```{r,echo=FALSE ,eval=FALSE,include=FALSE}
##16-retweet-network.R
```


```{r, include=FALSE}
rm(list =ls())
set.seed(100)
source("Head_file.R")
source("function.R")
options(digits=3)
includeComments = FALSE
# load("./0102/result_Android.RData")
# allTweets = FALSE

load("./0102/result_lambda.RData")
#load("./0102/result.RData")
allTweets = TRUE
```
### Data description:  

We sampled 100k out of 12M trump's followers As of October 10, 2016. Out of those followers, we can download the profile infomation and their friend list of about 76k followers and cannot access the others due to private setting or accounts closure.  Besides, we downloaded their tweets from their timeline and collected 1200 most recent tweets (due to time limit, we can collect up to 3200 in thoery).  Out those tweets, we extracted those are retweets/replies to @realDonaldTrump  excluded the tweets outside the time interval 2015-01-01 and 2016-11-08(included). 

This Analysis is about replies/retweets from  Trump's followers to Trump's tweets. We have totally `r nrow(el)` replies/retweets from Trump's `r length(unique(el[,1]))` followers to `r length(unique(el[,2]))` of Trump tweets.  Due deletion or other problem, we can have full access to `r nrow(tweets)` of those tweets, have id of the tweets they interacted with. 


### Goals and Conclusions
1, clustering based on the interaction between followers and Trump's tweets reveals that: 1, They are clearly a group tweets during a common time period, This indicates one interesting thing : different groups of  followers are engaging with Trump just during during different time period. Only very few people are engaging with him all the way through. the tweets don't seem to have a common theme in each clusters. 

2, clustering Trump's tweets using bag of words analysis reveal some common themes in each clusters -- Themes includes 

3, cluster-level analysis, combined with the analysis from who-following-who network.  Findings: 



```{r, include= includeComments}
## A: bipartie graph 
dim(A);dim(tweets); dim(users)
k = 10 
#users is recently downloadable followers among 8671
# tweets is  recently downloadable followers among 3741(2400) + tweets found in Hadoop(~930)
users_A <- users[match(rownames(A), users$id_str), ]; tweets_A <- tweets[match(colnames(A), tweets$id_str),]
sum(is.na(users_A)); sum(is.na(tweets_A))
users_A$id_str <- user_ids; tweets_A$id_str<- tweets_ids
#tweets over time
dates  <- c("2015-07-01 00:00:00", "2015-10-01 00:00:00","2016-01-01 00:00:00", 
            "2016-02-01 00:00:00", "2016-03-01 00:00:00",
            "2016-04-01 00:00:00","2016-05-01 00:00:00","2016-06-01 00:00:00","2016-07-01 00:00:00",
            "2016-08-01 00:00:00","2016-09-01 00:00:00","2016-10-01 00:00:00","2016-11-01 00:00:00",
            "2016-11-09 00:00:00")
findInterval <- function(time_vec, dates){ ##<=
  res <- integer(length(time_vec)) # initialized as 0
  for (i in length(dates):1){
    res[time_vec<= dates[i]] <- i
  }
  return (res)
}

frequency_time <- function(time_vec, dates){
  counts <- integer(length(dates))
  ids <- findInterval(time_vec, dates)
  for( i in 1:length(counts)){counts[i] <- sum(ids == i)}
  return(counts)
}
counts <-frequency_time(tweets$created_at, dates)
p1 <- ggplot(data = data.frame(dates = gsub('(.*) .*', '\\1', x= dates),counts = counts),
       aes(x = dates, y = counts))+geom_bar(stat='identity')+
  ggtitle(("Trump's tweets counts over time"))+
     theme(axis.text.x = element_text(angle = 45, hjust = 1))



nr <- dim(A)[1]; nc <- dim(A)[2]
Dr <- rowSums(A); Dc <- colSums(A)
dat <- data.frame(Dr = Dr); rownames(dat) <- NULL

# row degree
classify <- function(x, intervals){
  pos <- rep(NA, length(x))
  for ( i in 1:(length(intervals)-1) )
    pos[x >= intervals[i]] <- i 
  return (pos)
}
intervals <- c(1,2,4,10,50,max(Dr))
counts <- table(classify(Dr,  intervals)  )
tweets.by.user <- data.frame(order = factor(1:length(intervals[-1])), count = as.vector(counts)); 
tweets.by.user$order <-factor(tweets.by.user$order, levels = 1:length(counts), labels =   paste0('[',intervals[1:(length(intervals)-1)],',',intervals[2:length(intervals)], ')'))
p2 <- ggplot(tweets.by.user, aes(order, count, fill = order))+
  geom_bar(stat="identity") + labs(title = "number of replies/retweets by each followers", x = '', ylab = "Number")

## column degree very skewed, take log
power <- table(Dc)
tmp <- data.frame(degree = log2(as.numeric(names(power))) , freq = as.vector(log2(power)) )
coefs <- lm(freq~degree, data =tmp)$coef
power_law.plot <- ggplot(data =tmp , aes(x=degree, y=freq))+geom_point()+
  geom_smooth(method = 'lm', colour = 'red', se= FALSE) + labs(title = paste0("a=",round(coefs[1],2),", b=",round(coefs[2],2)), x = "log2 degree", y = 'log2 freq')
#power_law.pl

k = 10
Z <- membershipM(bip.result$row); Y <- membershipM(bip.result$col)
blockM <- t(Z) %*% A %*% Y
balloon.plot(blockM, xlabel  = paste0(1:k,' (',colSums(Y),")"), ylabel = paste0(1:k,' (',colSums(Z), ')')) +  labs( title = paste(sum(blockM), "retweets/replies between", sum(Z), "follower and", sum(Y), "tweets"), y = paste0("clusters for followers"), x= paste("clusters for tweets"))
colnames(blockM) <- paste0(1:k,'-',colSums(Y)); rownames(blockM) <- paste0(1:k,'-',colSums(Z)) 
blockM
in_out_ration = diag(blockM)/(rowSums(blockM)-diag(blockM)); names(in_out_ration) <- NULL

#partition tweets
clustered_tweets_by_retweet <- NULL
for ( i in 1:k){
  data <- tweets_A[ bip.result$col == i, ]
  clustered_tweets_by_retweet <- rbind(clustered_tweets_by_retweet, data.frame(cluster_id = rep(i, nrow(data)), data))
}
rownames(clustered_tweets_by_retweet) <- 1:nrow(clustered_tweets_by_retweet)
clustered_tweets_by_retweet <- clustered_tweets_by_retweet[which(!is.na(clustered_tweets_by_retweet$created_at)), ]
write.csv(clustered_tweets_by_retweet, file = "./0102/retweeting/clustered_tweets_by_retweet_retweeting.csv", row.names = F)

selected_tweets <- NULL  # select some tweets close to center
for ( i in 1:k){
  score <- irlba_L$v[,1:k][bip.result$col == i,] %*% as.vector(km1$centers[i,])
  tweets_i <- tweets_A[bip.result$col == i,]
  score <- score * (!is.na(tweets_i$created_at))
  data <- tweets_i[order(-score)[1:5],]
  selected_tweets <- rbind(selected_tweets, data.frame(cluster_id = rep(i, nrow(data)), data))
}
names(selected_tweets)
rownames(selected_tweets) <- NULL
write.csv(selected_tweets, file = "./0102/retweeting/selected_tweets_retweeting.xls",  row.names = F)

#twitter clusters distributed over time, total 3377
counts_cluster <- matrix(0,  k ,length(dates))   #tweets in each clusters
for( i in 1:k){
  tw1 <- clustered_tweets_by_retweet[clustered_tweets_by_retweet$cluster_id == i, ]
  counts_cluster[i,] <- frequency_time(tw1$created_at, dates)
}
## two plots below show that retweeting pattern heavily depends on time

dat <- flattenMatrix(counts_cluster) ## by col
names(dat) <- c("cluster","time", "count");
dat$time <- as.factor(dat$time); 
dat$time <- factor(dat$time, 1:length(dates), labels = gsub('(.*) .*', '\\1', x= dates)) 
dat$cluster <- as.factor(dat$cluster)
p4 <- qplot(x = time, y = count, data =dat, group = cluster, geom = 'line', colour = cluster)+facet_grid(cluster~.)+
  ggtitle(paste("Trump's tweets frequency over time (",nrow(tweets),"0"))+
  theme(axis.text.x = element_text(angle = 45, hjust = 1))


p5 <- balloon.plot(counts_cluster, xlabel = gsub('(.*) .*', '\\1', x= dates),
                   ylabel = )+ labs(title = " different clusters of  Trump's tweets (total 1034, found in retweeting)  \n creted during 2015-01-01 and 2016-11-08") 
counts_cluster


if (allTweets == TRUE){
  labels_twitter_bip <- labels_twitter_bip <- c("Oct","earlyNov","Bef2016Oct+", "Bef2016","Mar+Apr","AugSep+","AprMay+", "MayJun+","JulAug+","July")
  expected.ord1 <- c(4,5,7,8,10,9,6,3,1,2)
}else{
  labels_twitter_bip <- labels_twitter_bip <- c("July","Aug","MayJun", "May","MarApr","Bef2016","Sep", "unif_small","OctNov","Oct")
expected.ord1 <- c(8,6,5,4,3,1,2,7,10,9 )
}


```





### Section 1,  Anaysis on retweeting bipartite graph on `r dim(A)[1]` followers and `r dim(A)[2]` tweets matrix. Clustering followers and tweets simultaneouly.


```{r}
dim(A)
p1   # exploring, trump's tweet over time
p2  # followers' engagement frequency, coldegree

## distribution of # of retweets
power_law.plot 

tmp <- cbind(retweet_time = colSums(A)[which(colSums(A) > 200)], tweets_A[which(colSums(A) > 200),1:3])
tmp

# clustering from retweeting 
blockM <- t(Z) %*% A %*% Y
balloon.plot(blockM, xlabel  = paste0(1:k,'-',colSums(Y)), ylabel = paste0(1:k,'-',colSums(Z))) +  
    labs( title = paste(sum(blockM), "retweets/replies between", sum(Z), "follower and", sum(Y), "tweets"), y = paste0("clusters for followers"), x= paste("clusters for tweets"))
colnames(blockM) <- paste0(1:k,'-',colSums(Y)); rownames(blockM) <- paste0(1:k,'-',colSums(Z)) 
blockM
in_out_ration = diag(blockM)/(rowSums(blockM)-diag(blockM)); names(in_out_ration) <- NULL
in_out_ration

#p4 ## line plot, clusters over time
p5 <- balloon.plot(counts_cluster, xlabel = gsub('(.*) .*', '\\1', x= dates),
                   ylabel = )+ labs(title = " different clusters of  Trump's tweets (total 1034, found in retweeting)  \n creted during 2015-01-01 and 2016-11-08") 

p5  ## ballplot 

counts_cluster
## cluster 8, very tiny, but contain the most frequnt tweets
#clustered_tweets_by_retweet[clustered_tweets_by_retweet$cluster_id==8,]$text

#most frequently retweeted tweets
tmp <- cbind(retweet_time = colSums(A)[which(colSums(A) > 200)], tweets_A[which(colSums(A) > 200),1:3])
tmp
```


#### Finding 1, the block structure is very clear, the ratio of the number retweets/replies within cluster over that outside clusters is `r in_out_ration`. 

#### Finding 2, The structures of reteweet/reply network clearly associated with time. During different time periods, different groups of followers were retweeting different groups of tweets. 


```{r, echo = includeComments}
## on topic modeling  #2694 words
k =20
U <- irlba_Llabda$u
Z <- membershipM(km3$cluster)
centers <-  Diagonal(k,colSums(Z)^(-1)) %*% t(Z) %*% At # centers in the original space
terms <- colnames(At); mean_vec <- colMeans(At)
pdf(paste0("./0102/combined/wordcloud_",k,".pdf"))
for (i in 1:k){
    diff = sqrt(centers[i,]) - sqrt(mean_vec)
    pos_idx <- which(diff > 0)
    wordcloud(words = terms[pos_idx], freq = diff[pos_idx], 
              max.words =50,
              rot.per = 0, random.order = F, scale = c(2,0.3))
    title(paste0("cluster of size: ", km3$size[i]))
}
dev.off() 
selected_words <- matrix("", 20, k)
for ( i in 1:k){
  diff = sqrt(centers[i,]) - sqrt(mean_vec)
  selected_words[,i] = terms[order(-diff)[1:20]]
  selected_words[1:10,]
}
clustered_tweets_by_combined <- NULL
for( i in 1:k){
    #i= 15
    tweets_i <- tweets[which( km3$cluster == i),]
    scores <- U[which(km3$cluster == i),1:k]  %*% matrix(km3$centers[i,])
    clustered_tweets_by_combined <- rbind( clustered_tweets_by_combined, 
                              data.frame(cluster_id = rep(i, nrow(tweets_i)),
                                      tweets_i[order(-scores),])  )                         
}
write.csv(clustered_tweets_by_combined, 
          file =paste0("./0102/combined/clustered_tweets_text_k",k,".csv"), row.names = F)

selected_tweets <- NULL
 for( i in 1:k){
    tweets_i <- tweets[which(km3$cluster == i),]
    scores <- U[which(km3$cluster == i),1:k]  %*% as.vector(km3$centers[i,])
    tweets_i <- tweets_i[order(-scores)[1:5],]
    selected_tweets<- rbind(selected_tweets, 
                              data.frame(cluster_id = rep(i, nrow(tweets_i)),tweets_i)  )                 }
rownames(selected_tweets) <- NULL
#selected_tweets$text
write.csv(selected_tweets, 
           file =paste0("./0102/combined/selected_tweets_text_k",k,".csv"),                          row.names = F)

# name the clusters
#8 -- Goofy Elizabeth Warren Our Native American Senator ,"failing @nytimes"
if(allTweets == TRUE){
labels_twitter_text <-  1:k
  
  #c("voteTrump_primary", "elizabethWarren","debates",
                         # "enjoyInterview","trump16_convention","#amerifirst_#imwithyou", 
                         # "campaign_join","urgeVote", "crooked_drainswamp",
                         # "law_police_condolence",
                         # "trumprally_joinme",
                         # "specialInterests","MAGA_#primary",
                         # "riggedSystem_email","tedcruz_rep",
                         # "nytime_cnn","urgeDonation",
                         # "clintion_sanders_obama",
                         # "job_border",
                         # "poll_debate_@megyn",)
expected.ord2 <- 1:k
  #c(4,17,9,3,14,15,5,19,6,11,13,18,1,2,12,7,8,10,16,20)
}else{
  labels_twitter_text <-
  c(
  "hillary_emails_rigged",
  "dishonest_media",
  "obama_isis_warming" ,
  "economy_jobs",
  "sanders_sold_rigged",
  "caimpaign_join",
  "cnn_morningjoe_fox",
  "citing_polls_news",
  "cruz_kasich",
  "MAGA_slogan",
  "convention_ryan",
  "wonPrimary_unfair",
  "terrorism_islam",
  "campaign_thankyou",
  "mitt_romney",
  "border_immigration",
  "failingnytimes",
  "specialInterests",
  "beingInterviewed",
  "elizabethWarren"
  )
expected.ord2 <- c(1,3,9,5,15,20,11,12,2,7,17,10,4,13,16,18,6,8,14,19)
}




## topics change over time 
counts_time_text <- matrix(0, length(dates), k)   # number of tweets in each clusters
for( i in 1:k){
  tw1 <- clustered_tweets_by_combined[clustered_tweets_by_combined$cluster_id == i, ]
  counts_time_text[,i] <- frequency_time(tw1$created_at, dates)
}

dat2 <- flattenMatrix(counts_time_text)
names(dat2) <- c("time","cluster", "count");
dat2$time <- as.factor(dat2$time); dat2$time <- factor(1:length(dates), labels = gsub('(.*) .*', '\\1', x= dates)); dat2$cluster <- as.factor(dat2$cluster)
p6 <- qplot(x = time, y = count,data =dat2, geom = "line", group = cluster, colour = cluster)+facet_grid(cluster~.)+ ggtitle("Trump's tweets frequency over time (2419)")+
  theme(axis.text.x = element_text(angle = 45, hjust = 1))
p7 <- balloon.plot(t(counts_time_text)[expected.ord2,], 
                   ylabel = paste0(labels_twitter_text[expected.ord2], '-',km3$size[expected.ord2]),
                   xlabel = gsub('(.*) .*', '\\1', x= dates))+ labs(title = paste0("different types of Trump's tweets (total ", sum(counts_time_text)," ) \n during 2015-01-01 and 2016-11-08"))
#p7

# I did a bigram on the text, which has the advantages to taking into account the order and create more words for our use. also I double the importance of the hashtags.

```


### Section 2, Clustering tweets based on their contents, bag of words analysis
From the read the tweets and look the frequent words of each clusters, we find there are follwoing themes

```{r}
mds_fit <- cmdscale(dist(km3$centers),eig=TRUE, k=3)
par(mfrow = c(2,2))
plot(mds_fit$points[,1:2], type = 'n', main="visualization of cluster centers", 
     xlab = "1st coordinate in MDS", ylab = "2nd coordinate in MDS" ) 
     text(mds_fit$points[,1:2], labels = 1:k, col = 1:k)
plot(mds_fit$points[,c(1,3)], type = 'n', main="visualization of cluster centers")
text(mds_fit$points[,c(1,3)], labels = 1:k,col = 1:k)
plot(mds_fit$points[,2:3], type = 'n' , main="visualization of cluster centers")
text(mds_fit$points[,2:3], labels = 1:k, col = 1:k)
barplot(km3$size, main="cluster size")
par(mfrow = c(1,1))
labels_twitter_text

selected_words 

# p7 <- balloon.plot(t(counts_time_text)[expected.ord2,], 
#                    ylabel = paste0(labels_twitter_text[expected.ord2],'-',km3$size[expected.ord2]),
#                    xlabel = gsub('(.*) .*', '\\1', x= dates))+ labs(title = "different types of Trump's tweets (total 1034) \n during 2015-01-01 and 2016-11-08")
p7

```








## Section 3, cluster level interaction
for each cluster of Trump followers,  what are the distributions of  types they are retweet/replies to

```{r, echo=F}
library(xlsx)
following_cluster <- read.csv("./1209/following/k50/id_sn_cluster.csv", colClasses =  c("character","character","integer"), stringsAsFactors = F)

cluster_category <- read.xlsx("Topics and Following clusters.xlsx", 
                       sheetName= 2, stringsAsFactors = F)
Z <- membershipM(following_cluster$cluster)
followingZ <- Z[match(user_ids, following_cluster$id_str),]; dim(followingZ)

cluster.name <- cluster_category$Name
cluster.name[9] <- "liberal media I";  cluster.name[12] <- "liberal media II"; 
cluster.name[20] <- "celebrity, sports I [probably millenials]"
cluster.name[20] <- "celebrity, sports II [probably millenials]"
if (length(which(nchar(cluster.name)>25))>0){
  cat("some cluster names are too long, get truncated at 25!!!")
  cluster.name <- substr(cluster.name, 1, 25)
}

colnames(followingZ) <- cluster.name 
csize <- c(which(!is.na(cluster_category$Category)), 51) #accumulated position
categoryZ <- matrix(0, 50, length(csize)-1)    # 50 cluster x 10 category
for(i in 1: (length(csize)-1)){
    tmp <- cluster_category$Cluster...1[csize[i]:(csize[i+1]-1)]
    categoryZ[tmp, i] <- 1
}
category.name  <- cluster_category$Category[csize[1:(length(csize)-1)]]

# reduced categry.names
category.name <- c("Trump supporters","conservatives","liberals","politically engaged","public affairs minded", "civic-minded and engaged","politically disengaged","tech savvy","regional","international" )
colnames(categoryZ) <- category.name
rownames(categoryZ)  <- cluster.name

##ordering the cluster.name
ord <- NULL
for(i in 1:dim(categoryZ)[2]){ord <- c(ord, which(categoryZ[,i] == 1))}
cluster.name <-cluster.name[ord]
categoryZ <- categoryZ[ord,]
followingZ <- followingZ[,ord]
Z <- Z[,ord]

##category level membership matrix
Z_followers_category <- followingZ %*% categoryZ ; 


tmp <- data.frame(clustername = unlist(lapply(cluster.name, function(x) substr(x, 1,30))),
           size_retweet = colSums(followingZ), size_in_76k = colSums(Z), ratio_of_engage =        round(colSums(followingZ)/colSums(Z),2))
rownames(tmp) <- NULL; tmp



timeY <- membershipM(findInterval(tweets_A$created_at, dates))
colnames(timeY) <- gsub('(.*) .*', '\\1', x= dates)


### retweet volumne, 50 clusters x 14 time periods of tweets over time
blockM <-t(followingZ  %*% Diagonal(dim(followingZ)[2], colSums(followingZ)^(-1))) %*%u2t %*% timeY
balloon.plot(blockM, ylabel = paste0(rownames(blockM), ' (', colSums(followingZ), ')'),
             xlabel = paste0(colnames(blockM),' (',colSums(timeY), ')')) + ggtitle( paste("Trump followers engagement with his tweets over time\n", nrow(el),"retweets between", sum(followingZ), "followers x", sum(timeY), "tweets: \n among 50 following vs  14 time periods") )
colnames(blockM) <- NULL; blockM



### retweet volumne, 10 categories x 14 time periods of tweets over time
blockM <-t(Z_followers_category  %*% Diagonal(dim(Z_followers_category)[2], colSums(Z_followers_category)^(-1))) %*%u2t %*% timeY
balloon.plot(blockM, ylabel = paste0(rownames(blockM), ' (', colSums(Z_followers_category), ')'),
             xlabel = paste0(colnames(blockM),' (',colSums(timeY) ,')')) + ggtitle( paste("Trump followers engagement with his tweets over time\n", nrow(el),"retweets between", sum(followingZ), "followers x", sum(timeY), "tweets: \n among 50 following vs  14 time periods") )
colnames(blockM) <- NULL; blockM

### combing with retweeting:
retweetingZ <- membershipM(bip.result$row);  
colnames(retweetingZ) <- labels_twitter_bip
# cat("size of followers cluster/tweets based on retweeting pattern 5569x1034:")
# tmp <- data.frame(clustername = labels_twitter_bip,size= colSums(retweetingZ) )
# rownames(tmp) <- NULL ; tmp
retweetingY <- membershipM(bip.result$col); 
colnames(retweetingY) <- labels_twitter_bip
# tmp <- data.frame(clustername = labels_twitter_bip, size = colSums(retweetingY))
# rownames(tmp) <- NULL; tmp
## reordering
ord <- expected.ord1 # time 
retweetingZ <- retweetingZ[,ord]; retweetingY <- retweetingY[,ord]

# blockM <-t(followingZ) %*% retweetingZ
# cat(" partition of",dim(followingZ)[1],"followers: \n 50 following vs  10 retweeting (interpreted as time)")
# balloon.plot(blockM, ylabel = paste0(rownames(blockM), '-', colSums(followingZ)),
#              xlabel = paste0(colnames(blockM),'-',colSums(retweetingZ))) + ggtitle(paste(" partition of",dim(followingZ)[1],"followers: \n 50 following vs  10 retweeting") )
# blockM
# 
# 
# 
# blockM <-t(followingZ) %*%u2t %*% retweetingY
# cat(sum(retweetingZ), "followers x", sum(retweetingY), "tweets: \n 50 following vs  10 retweeting")
# balloon.plot(blockM, ylabel = paste0(rownames(blockM), '-', colSums(followingZ)),
#              xlabel = paste0(colnames(blockM),'-',colSums(retweetingY))) + ggtitle( paste("partition of ", nrow(el),"retweets between", sum(retweetingZ), "followers x", sum(retweetingY), "tweets: \n among 50 following vs  10 retweeting") )
# blockM
# 
# 
# blockM <-t( followingZ %*% Diagonal(dim(followingZ)[2],colSums(followingZ)^(-1))) %*%u2t %*% retweetingY
# cat(sum(retweetingZ), "followers x", sum(retweetingY), "tweets: \n 50 following vs  10 retweeting")
# balloon.plot(blockM, ylabel = paste0(rownames(blockM), '-', colSums(followingZ)),
#              xlabel = paste0(colnames(blockM),'-',colSums(retweetingY))) + ggtitle( paste("partition of ", nrow(el),"retweets between", sum(retweetingZ), "followers x", sum(retweetingY), "tweets: \n among 50 following vs  10 retweeting") )
# blockM                 
# 
# 
# 
# #size of categories:
# # data.frame(categoryName = category.name, size = colSums(Z_followers_category), row.names = 1:length(category.name) )
# 
# 
# 
# blockM <- t(Z_followers_category) %*% retweetingZ  # 10 catergories of followers x 10 cluster of tweets
# balloon.plot(blockM, ylabel = paste0(rownames(blockM), '-', colSums(Z_followers_category)),  
#              xlabel = paste0(colnames(retweetingZ),'-',colSums(blockM)),
#              main = paste("partition of", sum(Z_followers_category)," followers: \n twitter clusters from bipartite \n how many followers in row cluster and column cluster?"))
# colnames(blockM) <- 1:ncol(blockM); blockM
# 
# 
# # how many followers in column cluster and column cluster
# blockM <- t(Z_followers_category) %*% u2t %*% retweetingY
# balloon.plot(blockM, ylabel = paste0(rownames(blockM), '-', colSums(Z_followers_category)),  
#              xlabel = paste0(colnames(retweetingY),'-',colSums(blockM)))+
#   ggtitle(paste(sum(retweetingZ), "followers x", sum(retweetingY), "tweets: \n 10 following vs  10 retweeting \n how many retwees happen in row cluster and column cluster?") )
# colnames(blockM) <- 1:ncol(blockM); blockM
# 
# 
# 
# #Each colum in t(blockM) is a type of followers           
# blockM <- t(Z_followers_category%*%Diagonal(dim(Z_followers_category)[2], colSums(Z_followers_category)^(-1))) %*% u2t %*% retweetingY 
# balloon.plot(blockM, ylabel = paste0(rownames(blockM), '-', colSums(Z_followers_category)),  
#              xlabel = paste0(colnames(blockM),'-',colSums(retweetingY)),
#              main = "#retweets of each categories(retweeting) by each follower \nfrom different categories")
# colnames(blockM) <- 1:ncol(blockM); blockM
# 

### exploring the follower clusters and tweet(topic) clusters.

textY <- membershipM(km3$cluster)
colnames(textY) <- labels_twitter_text
ord <- expected.ord2
textY <- textY[,ord];
#u2t <- graph_component(el, connected = F)

cat(" following clusters x twitter topics : 50 clusters ")
blockM <- t(followingZ %*% Diagonal(dim(followingZ)[2],colSums(followingZ)^(-1))) %*% u2t[,match(tweets$id_str,tweets_ids)] %*% textY
balloon.plot(blockM, ylabel = paste0(rownames(blockM),' (', colSums(followingZ), ')'),
             xlabel = paste0(colnames(blockM),' (',colSums(textY),')'),
             main = "#engatement(retweets/replies to Trump's tweets) by each follower \n from 50 different clusters of Trump followers ")
colnames(blockM) <- 1:ncol(blockM); blockM


cat(" following clusters x twitter topics : 10 categories weighted by row")
blockM <-  t(Z_followers_category %*% Diagonal(dim(Z_followers_category)[2], colSums(Z_followers_category)^(-1)))  %*% u2t[,match(tweets$id_str,tweets_ids)]%*% textY
plot.f2t <- balloon.plot(blockM, ylabel =  paste0(rownames(blockM), ' (', colSums(Z_followers_category), ')'),
             xlabel = paste0(colnames(blockM), ' (',colSums(textY), ')'),
             main = "#engatement(retweets/replies to Trump's tweets) by each follower \n from 10 different categories of Trump followers ")
colnames(blockM) <- 1:ncol(blockM); blockM
plot.f2t

pdf('./0102/followers_tweets-1.pdf', width = 8, height = 6)
plot.f2t
dev.off()

cat(" following clusters x twitter topics : 10 categories weighted by row & col")
blockM <-  t(Z_followers_category %*% Diagonal(dim(Z_followers_category)[2], colSums(Z_followers_category)^(-1)))  %*% u2t[,match(tweets$id_str,tweets_ids)]%*% textY %*% Diagonal(dim(textY)[2], colSums(textY)^(-1))
plot.f2t <- balloon.plot(blockM, ylabel =  paste0(rownames(blockM), ' (', colSums(Z_followers_category), ')'),
             xlabel = paste0(colnames(blockM), ' (',colSums(textY), ')'),
             main = "#engatement(retweets/replies to Trump's tweets) by each follower \n from 10 different categories of Trump followers ")
colnames(blockM) <- 1:ncol(blockM); blockM
plot.f2t
#save this plot
pdf('./0102/followers_tweets-2.pdf', width = 8, height = 6)
plot.f2t
dev.off()

```



### Section 4, Sentiment analysis. For different groups of Trump's followers, will you tend to retweet, replies positively to Trump' tweets?

```{r, echo = F}
samp_tweets <- read.csv("../data/trump_tweets/followers_tweets/samp_retweets_alllabeled.csv", colClasses = c("character"), stringsAsFactors = F)
names(samp_tweets)[5] <- 'LABEL'
#when file opened by excel, there will be rounding error to the id_str, string to numeric, cause rounding error
samp_tweets1 <- read.csv("../data/trump_tweets/followers_tweets/samp_retweets.txt", colClasses = c("character"), stringsAsFactors = F)
samp_tweets$user_id_str <- samp_tweets1$user_id_str
idx1 <- match(samp_tweets$user_id_str, users_A$id_str)
samp_tweets <- samp_tweets[!is.na(idx1),]
dim(samp_tweets);
table(samp_tweets$LABEL)
samp_tweets$LABEL[samp_tweets$LABEL == 't'] <- 'Y'
samp_tweets$LABEL[samp_tweets$LABEL == 'y'] <- 'Y'
samp_tweets$LABEL[samp_tweets$LABEL == 'b'] <- 'N'
samp_tweets$LABEL[samp_tweets$LABEL == 'n'] <- 'N'
samp_tweets$LABEL[samp_tweets$LABEL == 'u'] <- 'U'
samp_tweets$LABEL[samp_tweets$LABEL == ''] <- NA

table(samp_tweets$LABEL, exclude = NULL)

opinionTrump <- rep('0', nrow(users_A))
idx <- match(samp_tweets1$user_id_str, users_A$id_str)
opinionTrump[idx[!is.na(idx)]] <- samp_tweets$LABEL; table(opinionTrump,exclude = NULL)
opinionTrump[opinionTrump == 'Y'] <- 1;  opinionTrump[opinionTrump == 'N'] <- 2; 
opinionTrump[opinionTrump == 'U'] <- 3; 
opinionTrump[opinionTrump == '0'] <- 4; #retweet

opinionTrump <- as.integer(opinionTrump); 
table(opinionTrump)
opinionZ <- membershipM(opinionTrump)
colnames(opinionZ) <- c("YES","NO","Unsure","ReTweet")


# balloon.plot(t(opinionZ)%*% followingZ %*%Diagonal(dim(followingZ)[2], colSums(followingZ)^(-1)) , xlabel = colnames(opinionZ), ylabel = colnames(followingZ))+ggtitle("followers' sentiment analysis")

# idx <- which(opinionTrump <3)
# blockM <- t(opinionZ[idx,1:2])%*% followingZ[idx,]


blockM <- Diagonal(dim(followingZ)[2],colSums(followingZ)^(-1)) %*% t(followingZ) %*% opinionZ
balloon.plot(blockM, xlabel = paste0(colnames(opinionZ),'-',colSums(opinionZ)), ylabel = paste0(colnames(followingZ),'-',colSums(followingZ)))+ ggtitle("percentage of YES/NOT/ReTweets in every follower(row =1)\n in each followers cluster(row cluster) ")
rownames(blockM) <-  substr(rownames(blockM), 1, 25) 

colMeans(blockM)
blockM


blockM <- t(followingZ) %*% opinionZ %*% Diagonal(dim(opinionZ)[2],colSums(opinionZ)^(-1))
balloon.plot(blockM, xlabel = colnames(opinionZ), ylabel = colnames(followingZ))+ ggtitle("the percentages different clusters of followers \n contribute to YES/NOT/ReTweets")
blockM
# blockM <- t(opinionZ[idx,1:2])%*% followingZ[idx,] %*%Diagonal(dim(followingZ[idx,])[2], colSums(followingZ[idx,])^(-1))
# balloon.plot(t(blockM) , xlabel = colnames(opinionZ[,1:2]), ylabel = colnames(followingZ))+ggtitle("% of supporter aamong each group of followers")



```



### Section 5, repeating the analysis based on the follower cluster learned from text analysis on the their timeline.

```{r, include=FALSE, eval=FALSE}
following_cluster1 <- read.csv("./1209/L2/k50/sn_cluster.csv", colClasses =  c("character","integer"), stringsAsFactors = F)
cluster_category1 <- read.xlsx("Topics and Following clusters.xlsx", 
                       sheetName= 1, 
                       stringsAsFactors = F)  

opinionTrump1 <- opinionTrump ## all followers
opinionZ1 <- membershipM(opinionTrump1)
colnames(opinionZ1) <- c("YES", "NO","Unsure","Retweet")
idx <- match(users_A$screen_name, following_cluster1$screen_name);sum(!is.na(idx))
opinionZ1 <- opinionZ1[!is.na(idx), ]; dim(opinionZ1)
followingZ1 <- membershipM(following_cluster1$cluster)
idx <- match(users_A$screen_name, following_cluster1$screen_name)
followingZ1 <- followingZ1[idx[!is.na(idx)],]

dim(followingZ1)
dim(opinionZ1)

cluster.name <- cluster_category1$topic.meaning
cluster.name[c(17,43,44)] <- paste0(cluster.name[c(17,43,44)],'-',1:3)
cluster.name[c(33,49)] <-paste0(cluster.name[c(33,49)],'-',1:2)
cluster.name[c( 35 ,8, 14,29,31)] <- paste0(cluster.name[c( 35 ,8, 14,29,31)] ,'-',1:5)
cluster.name[c( 19 ,25, 40)] <- paste0(cluster.name[c( 19 ,25, 40)]  ,'-',1:3)
cluster.name[c( 36 ,47)] <- paste0(cluster.name[c( 36 ,47)]  ,'-',1:2)
cluster.name[c( 18,22,50)] <- paste0(cluster.name[c( 18,22,50)]   ,'-',1:3)
cluster.name[c(13,39)] <- paste0(cluster.name[c(13,39)], '-',1:2)
cluster.name[c(1,23)] <- paste0(cluster.name[c(1,23)], '-', 1:2)
   cbind(order(cluster.name), sort(cluster.name))
if ( length(which(nchar(cluster.name)>25)) > 0){
  cat("some cluster names are too long, get truncated at 25!!!")
  cluster.name <- substr(cluster.name, 1, 25)
}
   
colnames(followingZ1) <- cluster.name
csize <- c(which(!is.na(cluster_category1$Category)), 51) #accumulated position
categoryZ1 <- matrix(0, 50, length(csize)-1)    # 50 cluster x 10 category
for(i in 1: (length(csize)-1)){
    tmp <- cluster_category1$topic...1[csize[i]:(csize[i+1]-1)]
    categoryZ1[tmp, i] <- 1
}
category.name <- cluster_category1$Category[csize[-length(csize)]]
if (length(which(nchar(category.name)>25))>0){
  cat("some category names are too long, get truncated at 25!!!")
  category.name <- substr(category.name, 1, 25)
  }

colnames(categoryZ1) <- category.name
rownames(categoryZ1) <- cluster.name


categoryZ1 <- categoryZ1[,c(2,1,3:dim(categoryZ1)[2])]; 
ord <-NULL
for(i in 1:dim(categoryZ1)[2]){ord <- c(ord, which(categoryZ1[,i]==1))}
followingZ1 <- followingZ1[,ord]; categoryZ1 <- categoryZ1[ord,]



blockM <- t(followingZ1) %*% opinionZ1; 
balloon.plot( blockM , xlabel = colnames(opinionZ1), ylabel =colnames(followingZ1)) + ggtitle(paste0("sentiment of the ", dim(followingZ1)[2]," clusters found based bag of words"))


balloon.plot( t(categoryZ1)%*% t(followingZ1) %*% opinionZ1 , xlabel = colnames(opinionZ1), ylabel =colnames(categoryZ1)) + ggtitle(paste0("sentiment of the ", dim(categoryZ1)[2]," categories found based bag of words"))
t(categoryZ1)%*% t(followingZ1) %*% opinionZ1
```


```{r, echo =FALSE,  eval= FALSE}

##partioning followers based on bag words analysis
## partioning tweets based retweeting or based text
following_cluster <- read.csv("./1209/L2/k50/sn_cluster.csv", colClasses =  c("character","integer"), stringsAsFactors = F)

cluster_category <- read.xlsx("Topics and Following clusters.xlsx", 
                       sheetName= 1, 
                       stringsAsFactors = F)    

Z <- membershipM(following_cluster$cluster)
followingZ <- Z[match(samp_tweets1$user_id_str, following_cluster$id_str),]; dim(followingZ)


csize <- c(which(!is.na(cluster_category$Category)), 51) #accumulated position
categoryZ <- matrix(0, 50, length(csize)-1)    # 50 cluster x 10 category
for(i in 1: (length(csize)-1)){
    tmp <- cluster_category$Cluster...1[csize[i]:(csize[i+1]-1)]
    categoryZ[tmp, i] <- 1
}
category.name  <- cluster_category$Category[csize[1:(length(csize)-1)]]
if ( length(which(nchar(cluster.name)>25)>0) ){
  cat("some cluster names are too long, get truncated at 25!!!")
  category.name <- substr(cluster.name, 1, 25)
}

category.name <- c("Trump supporters","conservatives","liberals","politically engaged","public affairs minded", "civic-minded and engaged","politically disengaged","tech savvy","regional","international" )

colnames(categoryZ) <- category.name
rownames(categoryZ)  <- cluster.name


words_in_followingclusters_pdf <- function(At, label, output){
  Z <- membershipM( label )
  clust_size <- colSums(Z)
  k <- length(clust_size)
  words_by_cluster <- Diagonal(k, clust_size^(-1)) %*% t(Z) %*% At
  mean_vec <- colMeans(At)
  terms <- colnames(At)
  library(wordcloud)
  pdf(output, width = 8, height = 9)
  sizes <- colSums(Z)
  for ( i in 1:k){
    #i =1
    diff <- words_by_cluster[i,] - mean_vec
    wordcloud(words = terms, freq = diff, 
              max.words =50,
              random.order = F, 
              scale = c(3,0.4))
    title(paste0("cluster of size: ", sizes[i]))
  }
  dev.off()
}
#words_in_followingclusters_pdf(At, label = bip.result$col,  
#                               output <- "../data/trump_tweets/followers_tweets/words_in_followingcluster_bigram.pdf")
                

```
